refactor(middleware): clarify middleware types and flow
