+++ src/postgraphile/extendedFormatError.ts
@@ -1,5 +1,7 @@
 import { GraphQLError } from 'graphql/error'
 import { PostGraphile } from '../interfaces'
+import GraphQLErrorExtended = PostGraphile.GraphQLErrorExtended
+import GraphQLFormattedErrorExtended = PostGraphile.GraphQLFormattedErrorExtended
 import mixed = PostGraphile.mixed
 
 /**
@@ -48,21 +50,3 @@ export function extendedFormatError(error: GraphQLError, fields: Array<string>):
   }
 }
 
-export type GraphQLFormattedErrorExtended = {
-  // This is ugly, really I just want `string | void` but apparently TypeScript doesn't support that.
-  [s: string]: Array<GraphQLErrorLocation> | Array<string | number> | string | void,
-  message: string,
-  locations: Array<GraphQLErrorLocation> | void,
-  path: Array<string | number> | void,
-}
-
-export type GraphQLErrorLocation = {
-  line: number,
-  column: number,
-}
-
-export type GraphQLErrorExtended = GraphQLError & {
-  hint: string,
-  detail: string,
-  code: string,
-}
